// packages/renderer/src/api/generated/types.gen.ts
export type OauthAccount = {
    id: string;
    oauth_name: string;
    access_token: string;
    account_id: string;
    account_email: string;
    expires_at?: number | null;
    refresh_token?: string | null;
};
// ...
export type User = {
    id: string;
    email: string;
    name?: string | null;
    is_superuser?: boolean;
    is_active?: boolean;
    is_verified?: boolean;
    has_password?: boolean;
    roles?: Array<UserRole> | null;
    oauth_accounts?: Array<OauthAccount> | null;
};

/**
 * UserCreate
 */
export type UserCreate = {
    email: string;
    password: string;
    name?: string | null;
    issuperuser?: boolean;
    isactive?: boolean;
    isverified?: boolean;
};

/**
 * UserRole
 */
export type UserRole = {
    role_id: string;
    role_slug: string;
    role_name: string;
    assigned_at: string;
};